Допили проект

+++ Backend/DrawMap.WebAPI/DTO/HeatMap.cs
@@ -4,6 +4,6 @@ namespace DrawMap.WebAPI.DTO;
 
 public class HeatMap
 {
-    public int VisitFrequency { get; set; }
+    public float VisitFrequency { get; set; }
     public LocationDto Location { get; set; }
 }
